Containers: improve error messages to users
- send a message if image could not be built
- send a message if image have the wrong architecture
